pyproject.toml
[project]
name = "website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.54.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]
